% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Generate a knitr report for a specified analysis}
\usage{
build_report(path, data, filename = NULL, title = "Example", seed = 47209)
}
\arguments{
\item{path}{Character vector specifying the path of a duplicated analysis template, should the user choose to use this instead}

\item{data}{A list object containing all relevant data elements. Specific attributes are defined in the details.}

\item{filename}{A character string specifying the name of the file. Default NULL, which generates a file called "[curtime]_output.html" in the working directory.}

\item{title}{A character string setting the report title. Default "Example".}

\item{seed}{A random seed set at the beginning of the report. Default 47209}

\item{analysis}{A character string specifing the type of analysis desired. Current options are: "tsp".}
}
\value{
An object, as specified in the description of the chosen analysis
}
\description{
Generate a knitr report for a specified analysis
}
\details{
This is a generic function that passes a data object to an analysis template
of the user's choosing. The template checks the data object to ensure compatibility.
}

