% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff_report.R
\name{diff_report}
\alias{diff_report}
\title{Compare an edited report template to the standard report template}
\usage{
diff_report(analysis, file_path, submit_diff = FALSE)
}
\arguments{
\item{analysis}{A character string specifing the desired analysis template. Current options are: "tsp".}

\item{file_path}{A character string specifying the path of the edited template file.}

\item{submit_diff}{Logical indicating whether or not to upload results as anonymous gist}
}
\description{
Compare an edited report template to the standard report template
}
\details{
This function uses diffr and htmlwidgets to render a file difference
between an existing package template and a user-edited template. An HTML
file is rendered and saved which displays the highlighted differences in
the two files. Optionally, the user may choose to upload this result
as an anonymous gist. If submit_diff=TRUE, the file is automatically uploaded
and a URL to the file is provided.
}

