% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicate_report.R
\name{duplicate_report}
\alias{duplicate_report}
\title{Duplicate a report template into the working directory}
\usage{
duplicate_report(analysis)
}
\arguments{
\item{analysis}{A character string specifing the type of analysis desired. Current options are: "tsp".}
}
\value{
The file path to the duplicated file, should the user later desire to
supply it to "diff_report".
}
\description{
Duplicate a report template into the working directory
}
\details{
This function duplicates an existing report template should the
user decide to make edits to any portion of the standard template. The function
will pop-up a file chooser dialog box so that the user may select a directory
in which to create the duplicated file. We encourage the user to use the
"diff_report" command if they make changes to the original report.
}

