% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/empirical_controls.R
\name{empirical_controls}
\alias{empirical_controls}
\title{Empirical controls calculation function}
\usage{
empirical_controls(exprs, n = 25)
}
\arguments{
\item{exprs}{An input matrix of gene expression values}

\item{n}{In each quantile bin, n genes are picked from the top and bottom}
}
\value{
A matrix with pairs in rows and samples in columns
}
\description{
This function calculates pairs for a data
set based on the empirical controls concept.
}

