% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.R
\name{fstats}
\alias{fstats}
\title{Calculate f-statistics quickly}
\usage{
fstats(dat, mod, mod0 = NULL)
}
\arguments{
\item{dat}{An input data set}

\item{mod}{The model matrix for the alternative model to be fit}

\item{mod0}{The model matrix for the null model to be fit. If NULL, the null model
is the model with only an intercept.}
}
\value{
fstats A vector of f-statistics
}
\description{
Using matrix algebbra, quickly calculate
f statistics for comparing nested models
}

