% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_pairs}
\alias{make_pairs}
\title{Utility/convenience functions}
\usage{
make_pairs(rn_min, rn_max, mat, n)
}
\arguments{
\item{rn_min}{Rownames for group 1}

\item{rn_max}{Rownames for group 2}

\item{mat}{Expression matrix (NOTE: must have rownames)}

\item{n}{Length of rn_min and rn_max}
}
\value{
A matrix of dimension n*n with pairs generated for
every comparision in rn_min to rn_max.
}
\description{
Make pairs out of two vectors of rownames (convenience function
to be used with empirical_controls.R)
}
\details{
This function creates a pair matrix using specifically-provided row names.
}

