% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_fs.R
\name{reg_fs}
\alias{reg_fs}
\title{Regression-based feature selection}
\usage{
reg_fs(pairmat, outcome, covar = NULL, npair = 5)
}
\arguments{
\item{pairmat}{An m x n matrix of pairwise features with rows=features, columns=samples}

\item{outcome}{A vector of outcomes of length n}

\item{covar}{An optional n x p matrix of additional covariates to adjust for}

\item{npair}{The number of paris we wish to select}
}
\value{
pairs A vector with the index for each chosen pair
}
\description{
Leverages the fact that the F-statistic for y~x and
x~y is the same to quickly choose pairs that add conditional
value for predicting an outcome. Also runs a separate pair-on-pair
regression to eliminate collinear pairs.
}

