% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sig_precision_crude.R
\name{sig_precision_crude}
\alias{sig_precision_crude}
\title{Crude estimate of precision gain due to gene signature predictions}
\usage{
sig_precision_crude(test_outcome, test_pred, test_covar)
}
\arguments{
\item{test_outcome}{A vector of length n outcomes for the test/validation set. Currently only supports binary outcomes.}

\item{test_pred}{Prediction vector of length n created using the decision tree model built by tsp_model_builder.R}

\item{test_covar}{A n x p matrix of additional covariates to adjust for. (optional)}
}
\value{
A crude approximation of the reduction in variance of a treatment effect estimator
for the outcome of interest.
}
\description{
This function takes training/test data and pairs generated via empirical control feature
selection and builds a decision tree model. It also cross-validates to get an out-of-sample
accuracy estimate
}
\details{
This function approximates how much precision gain we might expect
in a clinical trial setting if we adjused for predictions with our gene signature.
We can optionally provide additional covariates and estimate how much additional
gain we might observe if we include predictions from our model.
}

