% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{single_pairs}
\alias{single_pairs}
\title{Build individual pairs based on rownames}
\usage{
single_pairs(pair, mat)
}
\arguments{
\item{pair}{A single pairname "A<B", where "A" and "B" are valid rownames.}

\item{mat}{A matrix with rownames that contain all "A<B" pairs.}
}
\value{
A vector of 0's and 1's corresponding to the pair "A<B" we want
}
\description{
A convenience function to quickly build a small matrix of pairs. Used
when we want to predict on a validation set.
}

