% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tspreg_report.R
\name{tspreg_report}
\alias{tspreg_report}
\title{Generate a knitr report for model-building using tspreg}
\usage{
tspreg_report(data, outcome, covar = NULL, val = NULL, val_outcome = NULL,
  val_covar = NULL, npair = 5, filepath = NULL, title = "Example",
  seed = 47209)
}
\arguments{
\item{data}{A matrix of continuous data with features in rows
and samples in columns}

\item{outcome}{A binary outcome vector with length = ncol(data)}

\item{covar}{(optional) A design matrix of additional covariates to adjust for with samples in rows and covariates in columns}

\item{val}{(optional) A validation dataset, where feature rownames are contained in the feature rownames of the training data}

\item{val_outcome}{(optional) Outcome vector for validation data. Required if val is provided.}

\item{val_covar}{(optional) A design matrix of additional covariates. Required if covar and val are provided.}

\item{npair}{The number of pairs desired in the final model}

\item{filepath}{A character string of where the report should be generated. Default NULL, which generates a file called "[curtime]_output.html" in the working directory.}

\item{title}{A character string setting the rport title. Default "Example".}

\item{seed}{A random seed set at the beginning of the report. Default 47209}
}
\value{
The final decision tree model fit in the procedure.
}
\description{
Generate a knitr report for model-building using tspreg
}
\details{
This is a wrapper to the TSP model-building procedure. Minimal
input is a matrix of continous data from which to build features (i.e. a
gene expression matrix) and a binary outcome vector. This function will
supply the provided information to an R markdown template that runs
a standardized model-building procedure. The output will appear in the
working directory as an HTML report.

You may specify a full filepath to an output file. Just make sure the file has a
".html" extension.
}

